Adding initializations strategies tsp and clustering so that greedy, greedytopk, greedyrandomized

+++ cpp_generation/FIBI/src/algorithm/swaps/abstract.hpp
@@ -5,5 +5,5 @@ class SwapChoice
 {
 public:
 	SwapChoice() {};
-	virtual map<string, int> get_dict() = 0;
+	virtual map<string,long> get_dict() = 0;
 };
