+++ cpp_generation/FIBI/src/algorithm/swaps/maxsat.hpp
@@ -9,9 +9,9 @@ public:
 	var_id_t var_id;
 	MAXSATSwap() : var_id(-1) {};
 	MAXSATSwap(var_id_t var_id) : var_id(var_id), SwapChoice() {};
-	map<string, int> get_dict()
+	map<string,long> get_dict()
 	{
-		map<string, int> dict;
+		map<string,long> dict;
 		dict["var_id"] = var_id;
 		return dict;
 	};
