+++ cpp_generation/FIBI/src/algorithm/swaps/tsp.hpp
@@ -11,9 +11,9 @@ public:
 	town_in_tour_id_t j;
 	TSPSwap() : i(-1), j(-1) {};
 	TSPSwap(town_in_tour_id_t i, town_in_tour_id_t j) : i(i), j(j), SwapChoice() {};
-	map<string, int> get_dict()
+	map<string,long> get_dict()
 	{
-		map<string, int> dict;
+		map<string,long> dict;
 		dict["i"] = i;
 		dict["j"] = j;
 		return dict;
